Updated: configuration, validators, dto fields

+++ Server/Entities/DTOs/V1/Landmarks/LandmarkCreateDTO.cs
@@ -7,8 +7,8 @@ namespace Entities.DTOs.V1.Landmarks
         public ELandmarkType Type { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Latitude { get; set; }
-        public int Longitude { get; set; }
-        public int Altitude { get; set; }
+        public float Latitude { get; set; }
+        public float Longitude { get; set; }
+        public float Altitude { get; set; }
     }
 }
