+++ Server/Entities/Models/Landmarks/Landmark.cs
@@ -13,9 +13,9 @@ namespace Entities.Models.Landmarks
         public ELandmarkType Type { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Latitude { get; set; }
-        public int Longitude { get; set; }
-        public int Altitude { get; set; }
+        public float Latitude { get; set; }
+        public float Longitude { get; set; }
+        public float Altitude { get; set; }
 
         [Display(Name = "Created At Date")]
         public DateTime CreatedDate { get; set; }
